move the actual ticket and payment creation to the payment url processors (/pay/gocardless-start and /pay/transfer-start)
Redirct to the gocardless pay url rather than have the user click on it.

Stash our reference in the gocardless state thing when doing the gocardless
payments, get it out again when we get redirected back to so we can match it
to the gocardless resource_id and then store the gocardless resource_id
in the Payment object so later on we can match it on the webhook.

In the webhook lookup the gocardless resource_id (now just called id) and
find our payment object, then find the associated tickets and mark them as
paid along with payment.state. Thats the bare minimum of issue #5 done,
need to do something with the other bill actions, esp. failed.

handle gocardless.exceptions.SignatureError as well.

update the reconcilation stuff (issue #10 ) to match these changes, bank
transfer payments now mark tickets as paid and update payment.state -> paid.