Revert bank reference changes
This is something that needs to be stable as soon as we go live.  For this
reason only, I've left the bank reference on the Payment object instead of
the user (in case there's some need to identify the specific tickets).

I still think it'll confuse/inconvenience people who need to make multiple
payments, including all pre-payers.  A bank reference is almost always the
account number, and online banking doesn't make it easy to change for each
transaction.  Fortunately, we can enforce stability later.

I've removed the payment ID as well, because it makes the matching brittle
and isn't useful to the user.  A variable-length suffix is far too easy to
misread or incorrectly copy-paste, and with ~1000 payments in the database
a 4-digit token won't disambiguate enough.  (Not to mention that there's a
150% probability of collision on the token.)

Brief reasoning for other reverts:
- It's not a token, we don't care if somebody guesses another's reference
- Tokens are cryptographic, so you can't use random.sample for them anyway
- Structured references encourage people to interpret (i.e. misread) them
- We don't need EMF in the reference, as bank statements include payee
- Hyphens are a presentation issue, and should be ignored for matching
- We don't encode anything else with these bank-specific safechars
- Bank reference for a ticket is irrelevant to people using GoCardless
- What we're building is more a reservation system than a shopping cart