Separate creation of tickets from get_basket
The 'get_basket' function is used in a several places when either the
currently selected ticket types or the current cost are needed. This
causes problems as it also creates new Ticket entries. These tickets
should be created as part of payment processing once all information
(e.g. user id) is known.

To fix this problem this commit splits the 'get_basket' method in two.
'get_basket' which is a the general utility function and
'process_basket' which actually adds the tickets for the user. This
should cut down on problems and make the state of the db easier to
reason about.