Placeholder columns for versioning
Once we start the anonymisation process we want to work against specific
versions of a proposal. To do this we'll link forward i.e. when a
revision is made the old version will receive the id of the new version.
This means that you can find the most recent version of a submission by
looking for NULL.