Move handling of exceptions from cancel() into app
There are two cases when the cancel() method on payments may raise an
exception. Either because the payment has already been cancelled or if
the payment is via Stripe (which we cannot cancel). In both of these
cases handle the exception at the app where it can be logged and the
user alerted (rather than in the method or not at all as was the
previous case).