Add select field+checkbox for installation funds
Installations are most likely to requet funds and most likely to need
them early. They're also most likely to request large amounts so we
should capture this information as part of the submission process. To
force the decision (and hopefully reduce speculative/frivolous
requests) a checkbox must first be checked.