BUG FIX Remove unnecessary permissions check
The unnecessary checking of whether a user was a reviewer in certain cfp
pages would cause that user to receive a 403 if they have the 'reviewer'
permission and admin set (it is also assumed that this would be true for
anonymiser). This check is uneeded as no admin should be acting as
reviewers.