abstract and description should not be redundant
typically they are appended by consuming clients. 

For the sake of completeness:
Frab defines 'abstract' as  "One short paragraph that succinctly describes the event." and 'description' as "A longer description of the event. Use this to give additional details that did not fit in the abstract. Both, abstract and description, will be shown on the conference website. (HTML)"