Replace asset pipeline
This changeset replaces the Python/Flask/Webassets static asset pipeline
with JS/Gulp and over 10,000 npm packages.

Webassets is not especially well mantained these days, and it was at a
point where it would hinder us from writing modern JS/CSS. Given that
the majority of modern tooling for this is written in JS, a JS pipeline
makes more sense.

Specifically, this changeset does the following:
* Removes flask-webassets and webassets and replaces the Flask side of
  the operation with the pleasingly minimal
  [flask-static-digest](https://github.com/nickjj/flask-static-digest)
* Adopts Gulp as a build tool for static assets (my first choice was
  webpack but it's less well suited to non-JS-first sites)
* Sets up a JS build pipeline using Browserify/Babel/UglifyJS
* Sets up a CSS build pipeline using PostCSS (with postcss-preset-env
  which includes autoprefixer)/cleancss
* Moves JS and CSS source directories up to the root to avoid confusion
* Imports jQuery and Bootstrap 3 (CSS and JS) through npm modules and
  inserts them directly into main.js - this gives us more leeway for CSS
  customisation while reducing external dependencies
* Adds a docker image to run gulp in development
* Changes the production Dockerfile to compile static hashes and run the
  flask-static-digest compile stage

The schedule browser JS is almost certainly not working but we're aiming
to re-write that and it's difficult to test so I'm letting that fall by
the wayside for the moment.