Preserve filter order when updating proposals
It is often useful to work through specific subsets of CfP proposals.
Previously the order was very simply determined: filter by state then
order by modified and find the next item.

This retains the query string used to order proposals in the list view,
find the current proposal then sets the next. Whilst this is not
particularly elegant the number of submissions and how often this view
is used shouldn't cause undue problems.

When no filtering/sorting is set the previous method is used.