Bulk refund functionality, part 1
This refactors a lot of the refund logic in order to support the EMF
2020 cancellation refunds. It adds a `flask payments bulkrefund` task,
initially only supporting Stripe, which will process all pending refund
requests.

It does not attempt to resolve business-as-usual refunds - specifically
part-refunds - which are in a bit of a state. This is covered by #900
and needs re-visiting pre-2022.

This also adds an end-to-end Stripe flow test and so closes #162.

May also resolve parts of #753.