Require admission ticket for purchasing other items
We add a "requires_admission_ticket" attribute, which defaults to True,
indicating that the user must have an admission ticket (either in
their basket or already purchased) in order to purchase the product.

This prevents users from accidentally buying a parking or live-in
vehicle ticket without having an admission ticket.

If the user doesn't have an admission ticket, an error is shown on
the payment page, preventing them from completing the purchase.

Fixes #631