Fix log colourisation
As we're running under docker, we usually don't have a TTY.
You can now set COLORIZE_LOGS=always instead.

To show coloured logs on a system with journald, use --output=cat.

Don't enable colourisation in prod, because we also ship to syslog,
and don't want to fill that with escape sequences.