Add form to change a proposal's owner
We often end up creating submissions for invited users. At some point
it's useful to change the ownership of these back to the correct people
so we can communicate with them via the site/keep them updated on the
status of their talk slot/send them speaker tickets

(or in some rare cases they just used the wrong email address etc).

This adds a new form which makes that change.