Add 'get_days_with_slots' for lightning talks
Check that the day isn't in the past. This mostly replaces
'get_remaining_lightning_slots' (as it is mainly used for truthy-tests:
is this a valid day). It is kept in the admin view because there we want
to be able to inspect the state of the past

This excludes the case when the slot is for today because the greenroom
needs time to collage the slides etc.