Experiment: remove setuptools dependency from poetry.lock file
This is an attempt to fix a failing GitHub Actions pull request test workflow.

The failure manifests as "ModuleNotFoundError: No module named 'pkg_resources'" -- an attempt to import a setuptools module.

During the build logs, poetry removes setuptools, despite it clearly being added to poetry.lock by dependabot in the commit prior to this one.

```
- Removing setuptools (65.3.0)
- Updating mako (1.2.0 -> 1.2.2)
```

Based on some searching around in open python-poetry/poetry.git issues, removing the dependency from the lockfile apparently helps, and the root cause is a dependency filtering change between versions 1.1.x and 1.2.x of poetry.