Add 'create_tags' cfp task
We want the cfp tags to be relatively limited so it (should) be fine to
manage them via CLI. If need-be a management page can be added but I'd
like to avoid that for the time being as hopefully they'll be a
one-and-done set up. It's a separate command but I've also hooked it
into the `dev data` command just to make sure there's something there.