Switch from pyppeteer to playwright.
Fixes #1262.

This doesn't really improve a whole lot. playwright notably disables the
Chromium sandbox by default, so we don't need to do that explicitly any more,
but other than that the code looks largely the same.

The generated PDFs look slightly different, I think because playwright
install-deps installs more fonts, and we then end up using Arial for the
receipts.

Because this is substantially invasive and in the critical path for generating
receipts I don't think this is worthwhile merging for 2024.