Refactor refund form & stripe views
Move the logic of processing a refund/initialising the form in to
refundform class so it can be shared into other views (i.e. the soon to
be made self-refund view).

To make this work we want to separate out some of the logic from the
previous 'stripe' file so that the views are no longer entangled with
the various stripe process functions (and remove some circular
dependencies).