cfp: add video_download_url attribute to cfp
This is used to store a canonical url for the video recording of a CFP
item, used to automate release of the recording by means of the schedule
frab/json.
This is useful in particular when releasing videos via media.ccc.de
which can import our frab schedule and automatically download the videos
from the listed URLs, but it could also be used to automate youtube
uploads, etc.
We also add the ability to set this via the proposal api endpoint.