Add the proposal attribute table
There's a load of stuff that is only needed on some proposals (e.g.
capacity for workshop proposals, size for installations). Whilst these
will almost all get copied, 1:1 into schedule attributes if the proposal
is accepted it's going to be a lot easier to keep them in their own
table