Make secondary nav targets more clickable
At the moment, the hit target is just the link text itself. Instead, we
make the entire list element the click target, which makes it
easier to hit.

On mobile this doesn't matter anyway, because we collapse the nav into a
list of elements at the top of the page, so this only really affects
desktop.