Migrate from gulp to rsbuild
This substantially reduces the weight of the JS dependencies we pull in
via package.json.

rsbuild (https://rsbuild.rs) goes nyoom - this migrates the JS/TS
compilation from Babel to SWC, among other things, which helps
compilation go much faster.

To accommodate this, the "dist" directory is handed over to rsbuild, and
Flask is configured to serve static content from dist/static/ rather
than the root "static" directory, which is checked into Git. This is
because rsbuild is a bit smarter about trying to collect all the used
assets. Eventually it would be nice to stop copying _all_ of the assets
from e.g. octicons, and just copy the ones we need.