Correct session adding logic
Since cascade_backrefs was removed in 2.0, some of these objects were
not being added to the session.

In general, we should call session.add whenever (and only when) a new
object has been created. The cascade logic is finicky enough that it's
best not to rely on it in any way:

  https://gist.github.com/marksteward/051be3cd90ebf5f408ffeddca4c13a8f